Strukturerat om i koden. vänt på beroende mellan Peerson och Registrering

+++ loppservice/database/migrations/2023_11_02_183526_remove_one_more_failed_msrreg_for_oleg_volkov.php
@@ -4,6 +4,7 @@ use App\Models\Optional;
 use App\Models\Registration;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Schema;
 
@@ -14,12 +15,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Log::debug("Remove " . 'f2ae253d-7988-4e75-9e28-73353711febe');
-        $registration = Registration::find('f2ae253d-7988-4e75-9e28-73353711febe');
-        $registration->delete();
-        Optional::where('registration_uid', 'f2ae253d-7988-4e75-9e28-73353711febe')->delete();
+        if (App::isProduction()) {
+            Log::debug("Remove " . 'f2ae253d-7988-4e75-9e28-73353711febe');
+            $registration = Registration::find('f2ae253d-7988-4e75-9e28-73353711febe');
+            $registration->delete();
+            Optional::where('registration_uid', 'f2ae253d-7988-4e75-9e28-73353711febe')->delete();
 
-
+        }
     }
 
     /**
